requirements.txt
openai==1.8.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
